[chore] Moved fetch summaries action to its own file
